// src/app/model/tiposQuirurgicos.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Protocolos } from './protocolos';

export interface TiposQuirurgicos { 
    complicaciones?: string;
    dieresis?: string;
    examen?: string;
    exploracion?: string;
    exposicion?: string;
    idTipo?: number;
    procedimiento?: string;
    protocolo?: Protocolos;
    sintesis?: string;
}